test(hfresh): consolidate the teardown tests and the MultiGet benchmark
The pin tests, the objects-bucket tests and the MultiGet benchmark each sat
in their own file. Fold the teardown and pin tests into torn_store_test.go,
which already covered the already-gone case, and the benchmark into
store_test.go next to the rest of the PostingStore tests.

Tests and benchmark are unchanged apart from the file-level comment now
describing the whole set.
