+++ adapters/repos/db/vector/hfresh/store_test.go
@@ -12,6 +12,7 @@
 package hfresh
 
 import (
+	"context"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -152,5 +153,53 @@ func TestStore(t *testing.T) {
 		require.Equal(t, 2, len(p))
 		require.Equal(t, v, p[0])
 		require.Equal(t, v2, p[1])
+	})
+}
+
+// A search fans MultiGet out over every selected centroid, so this is where
+// per-access bucket pinning is felt: the pin is taken against a store-wide
+// lock shared by every bucket on the shard. Run with -cpu to see it under
+// concurrency, which is how it is actually reached.
+func BenchmarkPostingStoreMultiGet(b *testing.B) {
+	ctx := context.Background()
+
+	const (
+		postingCount    = 64
+		vectorsPerPost  = 32
+		vectorDimension = 8
+	)
+
+	store := testinghelpers.NewDummyStore(b)
+	cfg := StoreConfig{MakeBucketOptions: lsmkv.MakeNoopBucketOptions}
+	shared, err := NewSharedBucket(store, "bench", cfg)
+	if err != nil {
+		b.Fatal(err)
+	}
+	postings, err := NewPostingStore(store, shared, NewMetrics(nil, "n/a", "n/a"), "bench", cfg)
+	if err != nil {
+		b.Fatal(err)
+	}
+
+	ids := make([]uint64, 0, postingCount)
+	for postingID := range uint64(postingCount) {
+		var posting Posting
+		for v := range uint64(vectorsPerPost) {
+			posting = posting.AddVector(NewVector(postingID*vectorsPerPost+v, 1,
+				make([]byte, vectorDimension)))
+		}
+		if err := postings.Put(ctx, postingID, posting); err != nil {
+			b.Fatal(err)
+		}
+		ids = append(ids, postingID)
+	}
+
+	b.ReportAllocs()
+	b.ResetTimer()
+	b.RunParallel(func(pb *testing.PB) {
+		for pb.Next() {
+			if _, err := postings.MultiGet(ctx, ids); err != nil {
+				b.Fatal(err)
+			}
+		}
 	})
 }
